Fixed hash creation for Jobs
The creation of hashes for Jobs depended on serialize which in turn makes use of typeinformation of the types to serialize. By passing int from script and parsing string values, the hash for the same Job could be different. Introduced a forced string conversion to normalize the values before the hash creation

Fixed space handling in crontab file parsing since it messed up when multiple spaces were used or the line started with a space instead directly with the command

Extended test suite to enable regression tests for changed hashes
Refactored tests to remove duplication